+++ app/admin/adminCtrl.js
@@ -26,22 +26,29 @@
                 vm.navState = {
                     reports: 'cp'
                 };
-                if ($routeParams.productId && (vm.isChplAdmin() || vm.isAcbAdmin()))  {
+
+                // base case
+                if (vm.isChplAdmin() || vm.isAcbAdmin()) {
                     vm.navState.screen = 'dpManagement';
+                } else if (vm.isAtlAdmin()) {
+                    vm.navState.screen = 'atlManagement';
+                }
+                if (!vm.isChplAdmin()) {
+                    vm.navState.dpManagement = 'upload';
+                } else {
+                    vm.navState.dpManagement = 'manage';
+                }
+
+                // chosen section
+                if ($routeParams.section) {
+                    vm.navState.screen = $routeParams.section;
+                }
+                if ($routeParams.productId)  {
                     vm.navState.dpManagement = 'manage';
                     vm.productId = $routeParams.productId;
-                } else {
-                    if (vm.isChplAdmin() || vm.isAcbAdmin()) {
-                        vm.navState.screen = 'dpManagement';
-                    } else if (vm.isAtlAdmin()) {
-                        vm.navState.screen = 'atlManagement';
-                    }
-                    if (!vm.isChplAdmin()) {
-                        vm.navState.dpManagement = 'upload';
-                    } else {
-                        vm.navState.dpManagement = 'manage';
-                    }
                 }
+
+                // load editable acbs & atls
                 commonService.getAcbs(true, vm.isChplAdmin())
                     .then (function (data) {
                         vm.acbs = $filter('orderBy')(data.acbs,'name');