+++ app/admin/adminModule.js
@@ -3,12 +3,7 @@
 
     angular.module('app.admin', ['ngRoute', 'smart-table', 'app.common', 'app.loginServices', 'ngIdle', 'ngSanitize', 'ui.bootstrap'])
         .config(['$routeProvider', function($routeProvider) {
-            $routeProvider.when('/admin', {
-                templateUrl: 'admin/admin.html'
-            });
-        }])
-        .config(['$routeProvider', function($routeProvider) {
-            $routeProvider.when('/admin/:productId', {
+            $routeProvider.when('/admin/:section?/:productId?', {
                 templateUrl: 'admin/admin.html'
             });
         }])